+++ R/load_asset_data.R
@@ -43,6 +43,12 @@ load_asset_data <- function(path_terminal, trade_log_file, time_period = 1, data
               silent = TRUE)
   if(class(DFT1)[1] == "try-error") {stop("Error reading file. File with trades may not exist yet!",
                                        call. = FALSE)}
+  #add one column filled with zeroes   DFT1$X3 <- 0
+  #detect if some columns are filled with zeroes...
+  Z_detect <- lapply(DFT1, function(x) all(x == 0)) %>% as.data.frame()
+  #evaluate results, if any values are equal to 0 provide a warning
+  if(any(Z_detect == TRUE)){warning("Warning, one or more columns in the datafile contains zeroes")}
+
   if(!nrow(DFT1)==0){
     # data frame preparation
     DFT1$X1 <- ymd_hms(DFT1$X1)